+++ packages/genesys/packages/react-components-core/src/SelectorGroup/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Selector";
+export * from "./SelectorGroup.types";
+export * from "./SelectorGroupContext";